requirements.txt
loguru~=0.5.3
requests~=2.26.0
python-decouple
pyTelegramBotAPI
python-telegram-bot-calendar